Modif fonctionnel

+++ backend/src/main/java/fr/jfc/ptut/dto/SoignerForm.java
@@ -16,7 +16,7 @@ public class SoignerForm {
     private String utilisateur;
     private String unitduree;
     private String unitfreq;
-    private int doseparprise;
+    private double doseparprise;
     private int valduree;
     private int valfreq;
 }
